* rename the selector.ts to selectors.ts. * implement the getTempRecord and getTempValidRoutes selector.
