+++ src/index.ts
@@ -1,6 +1,6 @@
 // export { default as useTempData } from "./useTempData";
 // export { default as withTempData } from "./withTempData";
 export { default as tempDataReducer } from "./reducer";
-export { default as tempDataSelector } from "./selector";
+export { getTempData as tempDataSelector } from "./selectors";
 export { default as setupTempDataCleanup } from "./setupCleanup";
 export * from './types';
